shunt qs errors into phrases
